move randomizertypes to randomizer directory
