+++ soh/soh/Enhancements/randomizer/randomizer.h
@@ -5,7 +5,7 @@
 #include <string>
 #include "../../../include/ultra64.h"
 #include "../../../include/z64item.h"
-#include <randomizerTypes.h>
+#include "soh/Enhancements/randomizer/randomizerTypes.h"
 
 class Randomizer {
   private:
