+++ soh/soh/Enhancements/randomizer/randomizerTypes.h
